backend/pyproject.toml
[project]
name = "fake-currency-detection"
version = "1.0.0"
description = "Fake currency detection using MobileNetV3-Large CNN + OpenCV ensemble"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.10.0",
    "numpy>=1.26.0",
    "scikit-image>=0.24.0",
    "pillow>=10.4.0",
    "pymysql>=1.1.0",
    "cryptography>=43.0.0",
    "pytesseract>=0.3.13",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
]
